requirements.txt
# =============================================================================
# Construction Cost Estimator Agent - Requirements
# =============================================================================

# Core ADK and AI
google-adk>=1.2.1
google-genai>=1.19.0

# Environment and Configuration
python-dotenv>=1.0.0

# Web API (FastAPI)
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6

# Terminal Output
rich>=13.0.0

# Data Handling
pydantic>=2.0.0

# Async Support
aiohttp>=3.9.0

# PDF Generation
xhtml2pdf>=0.2.15

# Google AI
google-generativeai>=0.3.0

# Supabase Session Storage
supabase>=2.0.0
